inker image styling, random background colors
