+++ css/background.css
@@ -1,0 +1,14 @@
+.background-red {
+  background-color: #fa3c00; }
+
+.background-blue {
+  background-color: #3053CA; }
+
+.background-teal {
+  background-color: #5ccfc2; }
+
+.background-purple {
+  background-color: #a639b2; }
+
+.background-gray {
+  background-color: #939393; }
